compact by reducing reaction related headers to one file
